+++ pypesto/objective/constants.py
@@ -1,0 +1,13 @@
+"""
+This is for (string) constants used in the objective module.
+"""
+
+
+MODE_FUN = 'mode_fun'  # mode for function values
+MODE_RES = 'mode_res'  # mode for residuals
+FVAL = 'fval'  # function value
+GRAD = 'grad'  # gradient
+HESS = 'hess'  # Hessian
+HESSP = 'hessp'  # Hessian vector product
+RES = 'res'  # residual
+SRES = 'sres'  # residual sensitivities