+++ test/test_objective.py
@@ -10,20 +10,18 @@ import unittest
 
 class ObjectiveTest(unittest.TestCase):
 
-    def test_objective_separated(self):
-        obj = get_objective_rosen_separated()
-        self.check_evaluation(obj)
+    def test_evaluate(self):
+        for obj in [get_objective_rosen_separated(),
+                    get_objective_rosen_integrated()]:
+            self._test_evaluate(obj,
+                                rosen_x, rosen_fval_true,
+                                rosen_grad_true, rosen_hess_true)
+        self._test_evaluate(get_objective_poly_integrated(),
+                            poly_x, poly_fval_true,
+                            poly_grad_true, poly_hess_true)
 
-    def test_objective_integrated(self):
-        obj = get_objective_rosen_integrated()
-        self.check_evaluation(obj)
-
-    def check_evaluation(self, obj):
-        x = np.array([0, 1])
-
-        fval_true = 101.0
-        grad_true = [-2, 200]
-        hess_true = [[-398, 0], [0, 200]]
+    def _test_evaluate(self, obj,
+                       x, fval_true, grad_true, hess_true):
 
         # check function values
         fval, grad, hess = obj(x, (0, 1, 2))
@@ -44,6 +42,33 @@ class ObjectiveTest(unittest.TestCase):
         self.assertTrue(np.isclose(grad, grad_true).all())
         self.assertTrue(np.isclose(hess, hess_true).all())
 
+    def test_return_type(self):
+        for obj in [get_objective_rosen_separated(),
+                    get_objective_rosen_integrated()]:
+            self._test_return_type(obj,
+                                   rosen_x)
+        self._test_return_type(get_objective_poly_integrated(), poly_x)
+
+    def _test_return_type(self, obj,
+                          x):
+        ret = obj(x, (0,))
+        self.assertTrue(isinstance(ret, float))
+        ret = obj(x, (1,))
+        self.assertTrue(isinstance(ret, np.ndarray))
+        ret = obj(x, (2,))
+        self.assertTrue(isinstance(ret, np.ndarray))
+        ret = obj(x, (0, 1))
+        print(ret)
+        self.assertTrue(isinstance(ret, tuple))
+        self.assertTrue(len(ret) == 2)
+
+
+rosen_x = np.array([0., 1.])
+rosen_fval_true = 101.0
+rosen_grad_true = [-2., 200.]
+rosen_hess_true = [[-398., 0.],
+                   [0., 200.]]
+
 
 def get_objective_rosen_separated():
     return pypesto.Objective(fun=sp.optimize.rosen,
@@ -59,6 +84,21 @@ def get_objective_rosen_integrated():
     return pypesto.Objective(fun=rosenbrock,
                              grad=True,
                              hess=True)
+
+
+def get_objective_poly_integrated():
+    def poly(x):
+        return ((x - 2)**2 + 1,
+                2 * (x - 2),
+                2)
+    return pypesto.Objective(fun=poly, grad=True, hess=True)
+
+
+# for testing in 1d
+poly_x = 0.
+poly_fval_true = 5.
+poly_grad_true = -4
+poly_hess_true = 2
 
 
 if __name__ == '__main__':