+++ doc/releasenotes.rst
@@ -5,8 +5,14 @@ Release notes
 0.0 series
 ..........
 
+0.0.2 (2018-10-18)
+------------------
 
-0.0.1
------
+* Fix parameter values
+* Record trace of function values
+* Amici objective to directly handle amici models
 
-* Basic framework and implementation of the optimization.
+0.0.1 (2018-07-25)
+------------------
+
+* Basic framework and implementation of the optimization